requirements.txt
streamlit
google-api-python-client
youtube-transcript-api
requests
textblob
pandas
python-dotenv
